HMAN-985: Renamed error message field/key
Renamed errorMessage field on BadFutureHearingRequestException and HealthCheckException classes to errorResponse.  This makes it clearer what is being stored in the field.

Renamed apiErrorMessage key for health details on HearingManagementInterfaceHealthIndicator to apiErrorResponse.  Again, this is to make it clearer what the value of it is.

Updated tests to reflect change of name of field/key to error response.

Corrected a spelling mistake in WiremockFixtures.