HMAN-985: Include HMI endpoint health in hearing service health  (#618)
* HMAN-985: Added StandardCharsets import to build.gradle

Added StandardCharsets import to build.gradle.  This is needed to allow unqualified use of StandardCharsets.UTF_8 constant in loadEnvSecrets task.  The import appears to have been accidentally left out of the changes made for HMAN-865.

* HMAN-985: Include HMI endpoint health in hearing service health

Added HearingManagementInterfaceHealthIndicator. This adds a check of the Hearing Management Interface private health endpoint to the hearing service health output.

The hearing service health will be shown as down if the Hearing Management Interface private endpoint returns an error, or an error is encountered when attempting to connect to it. The hearing service health will also be shown as down if the call to Active Directory to get an access token returns an error, or an error is encountered when attempting to connect to it.

A retryer has been added to retry the connection to Active Directory. This is configurable via environment variables.

Cleaned up some tests and changed logging to use placeholders instead of string concatenation.

* HMAN-985: Renamed error message field/key

Renamed errorMessage field on BadFutureHearingRequestException and HealthCheckException classes to errorResponse.  This makes it clearer what is being stored in the field.

Renamed apiErrorMessage key for health details on HearingManagementInterfaceHealthIndicator to apiErrorResponse.  Again, this is to make it clearer what the value of it is.

Updated tests to reflect change of name of field/key to error response.

Corrected a spelling mistake in WiremockFixtures.

* HMAN-985: Removed fh-get-token-url from preview values

Removed fh-get-token-url entry from hmc key vault section of values.preview.template.yaml to see if it needs to be set there.

* HMAN-985: Added fh-get-token-url back to preview values

Added fh-get-token-url entry back to hmc key vault section of values.preview.template.yaml.  It is needed.

* HMAN-985: Use stubbed Active Directory API in preview

Changed values.preview.template.yaml to use stubbed Active Directory API from ccd-test-stubs-service rather than an actual Active Directory API.

* HMAN-985: Address review comments

Moved methods that create health status, AD API and HMI API test data to TestingUtil class.  Updated FutureHearingRepositoryIT, HearingManagementInterfaceHealthIndicatorIT and HearingManagementInterfaceHealthIndicatorTest to use methods added to TestingUtil class.

Removed a redundant List variable from hearingActualWithOutcomeEmpty() method in TestingUtil class.

* HMAN-985: Fixes following Spring Boot 3 update

- Replaced feign.client.config.future-hearing-api.readTimeout property in FutureHearingRepositoryIT and HearingManagementInterfaceHealthIndicatorIT classes with spring.cloud.openfeign.client.config.future-hearing-api.readTimeout property.
- FutureHearingRepositoryTest class: Cast null retryAfter parameter in feign RetryableException constructor to Long to avoid ambiguity with deprecated constructor which uses a Date for it.
- Replaced deprecated MockBean annotations in ManageExceptionsControllerTest class with MockitoBean annotations.

* HMAN-985: Changes following Spring Boot 3 update

Added JUnit5 named() method to first test data value created by adApiErrorsAndExpectedHealthCheckValues() and hmiApiErrorsAndExpectedHealthCheckValues() methods in TestingUtil class.  This allows parameterised tests to use the label defined by named() as the name of the test rather than displaying a list of all the parameters.

Changed parameterised tests in FutureHearingRepositoryIT and HearingManagementInterfaceHealthIndicatorIT that use the AD and HMI API error test data to display "named" label.